slight refactor, added utils alias
