+++ src/components/Navigation/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import { withRouter } from 'react-router';
-import tyr from '../../utils/tyr';
+import tyr from 'Utils/tyr';
 import './styles.scss';
 
 import Tear from '../common/Tear';
